some example "hello" programs
I have the idea of illustrating a few languages in small pieces
simultaneously.  The goal is really to just do simple abstraction -- the
kind of things that every language has:  functions and either direct or
indirect encodings of objects.

But I don't want to get all caught up in installing full-on projects for
these languages with complicated build systems, test frameworks, and
such.

Fortuntately, almost all the languages I'd want to teach in can be
written in executable, one-file scripts.  That means that so long as you
have the interpretter installed, you should be able to run these
programs.

So lets see if you can get the interpretters for:

    - Python (maybe already on your computer)
    - Ruby (similarly, you probably already have it)
    - Haskell (https://docs.haskellstack.org/en/stable/README/#how-to-install)
    - Scala (https://downloads.lightbend.com/scala/2.12.2/scala-2.12.2.tgz)

Then you should be able to run these programs by simply typing:

```
$ hello/hello-hs
$ hello/hello-js
$ hello/hello-py
$ hello/hello-scala
```