More Battleship
I was playing around with thinking about the problem in terms of types.
This is a really popular way to think about a problem when functional
programming in a language with a good type system and checker.

The next step is to define operations with type signatures.

The last thing to do is implement them.