rename some files
This is an old ops fixation of mine, but also connected to a pet peeve.
When we suffix a script as .sh or .py, we communicate this -- "this file
should be run by an interpretter directly, and not as an executable."

When we make a program executable, we should normally not have any
extension at all, because the "#!" line of the script declares which
interpretter we run on.  The whole point of this is to hide the
implementation from the user.  The user shouldn't care if their
executables are machine-compiled binary, or scripts run on some
interpretter.

So I'm naming these "functions-py" rather than "functions.py" because I
want it to be clear that you just call these directly as an executable.
I'm just putting the "-py" as a suffix because I might show latter how
to work with functions in other languages.  And if I work, say in
Haskell the ".hs" suffix will break down completely, because Haskell has
a different relationship with interpretation than Python.

I just thought I'd take a moment to explain these naming conventions,
because they connect with the deeper history of Unix systems.