A fully-fleshed out list implementation
Lot's of puzzles to keep you busy, and a few discussions I think are
good to go through along the way.

Also, I gitignored files ending in "-answered" to help me better develop
an answer key before scrubbing it to be an "-unanswered" file.  These
answer keys aren't committed for now.  Ideally, the answers you get are
more or less the same as the answer key -- especially after review.