Simplified "naive" representation
I didn't have to save the "head" and "tail" functions passed into the
constructors as "__head" and "__tail" fields only to call them as
methods later.  I can save them directly as "head" and "tail" fields,
and skip defining them as methods altogether.

This is more about Python, and less about the lesson.  It just
simplifies the code a touch.